moved edge indexes 1d / 2d to indexes_2d, and native_for_slim
