+++ autoarray/__init__.py
@@ -51,6 +51,7 @@ from .inversion.inversion.interferometer.mapping import InversionInterferometerM
 from .inversion.inversion.interferometer.lop import InversionInterferometerMappingPyLops
 from .inversion.linear_obj.linear_obj import LinearObj
 from .inversion.linear_obj.func_list import AbstractLinearObjFuncList
+from .mask.indexes_2d import Indexes2D
 from .mask.mask_1d import Mask1D
 from .mask.mask_2d import Mask2D
 from .operators.convolver import Convolver
