+++ test_autoarray/mask/test_mask_2d.py
@@ -564,22 +564,6 @@ def test__edge_buffed_mask():
     assert (mask.edge_buffed_mask == edge_buffed_mask_manual).all()
 
 
-def test__native_index_for_slim_index():
-
-    mask = aa.Mask2D.manual(
-        mask=[[True, True, True], [True, False, False], [True, True, False]],
-        pixel_scales=1.0,
-    )
-
-    sub_native_index_for_sub_slim_index_2d = (
-        aa.util.mask_2d.native_index_for_slim_index_2d_from(mask_2d=mask, sub_size=1)
-    )
-
-    assert mask.native_index_for_slim_index == pytest.approx(
-        sub_native_index_for_sub_slim_index_2d, 1e-4
-    )
-
-
 def test__unmasked_mask():
 
     mask = aa.Mask2D.manual(
@@ -670,30 +654,6 @@ def test__unmasked_1d_indexes():
     )
 
     assert mask.unmasked_1d_indexes == pytest.approx(unmasked_pixels_util, 1e-4)
-
-
-def test__edge_1d_2d_indexes():
-    mask = aa.Mask2D.manual(
-        mask=[
-            [True, True, True, True, True, True, True, True, True],
-            [True, False, False, False, False, False, False, False, True],
-            [True, False, True, True, True, True, True, False, True],
-            [True, False, True, False, False, False, True, False, True],
-            [True, False, True, False, True, False, True, False, True],
-            [True, False, True, False, False, False, True, False, True],
-            [True, False, True, True, True, True, True, False, True],
-            [True, False, False, False, False, False, False, False, True],
-            [True, True, True, True, True, True, True, True, True],
-        ],
-        pixel_scales=1.0,
-    )
-
-    edge_pixels_util = aa.util.mask_2d.edge_1d_indexes_from(mask_2d=mask)
-
-    assert mask.edge_1d_indexes == pytest.approx(edge_pixels_util, 1e-4)
-    assert mask.edge_2d_indexes[0] == pytest.approx(np.array([1, 1]), 1e-4)
-    assert mask.edge_2d_indexes[10] == pytest.approx(np.array([3, 3]), 1e-4)
-    assert mask.edge_1d_indexes.shape[0] == mask.edge_2d_indexes.shape[0]
 
 
 def test__edge_mask():
